Make random::RNG generic over RNG type
