+++ model/util/random.h
@@ -21,52 +21,136 @@
 #ifndef OM_util_random
 #define OM_util_random
 
+// Define to use boost distributions. Unfortunately these are not necessarily
+// value-stable or portable.
+// #define OM_RANDOM_USE_BOOST_DIST
+
 #include "Global.h"
+#include "util/errors.h"
 #include <set>
 #include <pcg_random.hpp>
 #include <boost/random/uniform_01.hpp>
 #include <gsl/gsl_rng.h>
 
+#ifdef OM_RANDOM_USE_BOOST_DIST
+#include <boost/random/normal_distribution.hpp>
+#include <boost/random/lognormal_distribution.hpp>
+#include <boost/random/gamma_distribution.hpp>
+#include <boost/random/beta_distribution.hpp>
+#include <boost/random/poisson_distribution.hpp>
+#include <boost/random/bernoulli_distribution.hpp>
+#include <boost/random/weibull_distribution.hpp>
+#endif
+
+#if !defined OM_RANDOM_USE_BOOST_DIST
+#include <gsl/gsl_cdf.h>
+#include <gsl/gsl_randist.h>
+#endif
+
+#include <cmath>
+
 namespace OM { namespace util {
 
+// Support functions
+template<class T>
+long unsigned int sample_ulong (void *ptr) {
+    T *rng = reinterpret_cast<T*>(ptr);
+    BOOST_STATIC_ASSERT (sizeof(uint32_t) <= sizeof(long unsigned int));
+    return static_cast<long unsigned int> ((*rng)());
+}
+template<class T>
+double sample_double01 (void *ptr) {
+    T *rng = reinterpret_cast<T*>(ptr);
+    boost::random::uniform_01<T&> dist (*rng);
+    return dist ();
+}
+
+template<class T>
+gsl_rng_type make_gsl_rng_type(T& rng) {
+    return {
+        "OM_RNG",		// name
+        rng.max(),		// max value
+        rng.min(),		// min value
+        0,				// size of state; not used here
+        nullptr,			// re-seed function; don't use
+        &sample_ulong<T>,
+        &sample_double01<T>
+    };
+}
+
 /// Our random number generator.
+template<class T>
 struct RNG {
     ///@brief Construction and checkpointing
     //@{
     /// Default constructor
-    RNG();
+    RNG(): m_rng(0) {
+        m_gsl_type = make_gsl_rng_type(m_rng);
+        m_gsl_gen.type = &m_gsl_type;
+        m_gsl_gen.state = reinterpret_cast<void*>(&m_rng);
+    }
     
     // Disable copying
     RNG(const RNG&) = delete;
     RNG& operator=(const RNG&) = delete;
   
     /// Allow moving, with explicit functions
-    RNG(RNG&& other);
-    void operator=(RNG&& other);
+    RNG(RNG&& other): m_rng(other.m_rng) {
+        m_gsl_type = other.m_gsl_type;
+        // The following point into instance of this RNG:
+        m_gsl_gen.type = &m_gsl_type;
+        m_gsl_gen.state = reinterpret_cast<void*>(&m_rng);
+    }
+    void operator=(RNG&& other) {
+        m_rng = other.m_rng;
+        m_gsl_type = other.m_gsl_type;
+        // The following point into instance of this RNG:
+        m_gsl_gen.type = &m_gsl_type;
+        m_gsl_gen.state = reinterpret_cast<void*>(&m_rng);
+    }
     
     /// Seed
-    void seed(uint64_t seed);
+    void seed(uint64_t seed) {
+        m_rng.seed(seed);
+    }
     
-    /// Write checkpoint
-    void checkpoint (ostream& stream);
-    /// Read checkpoint (note: this relies on the RNG having been constructed with the same seed)
-    void checkpoint (istream& stream);
+    /// Checkpointing
+    /// 
+    /// Note: this relies on the RNG having been constructed with the same seed.
+    template<class S>
+    void checkpoint(S& stream) {
+        m_rng.binary_checkpoint(stream);
+    }
     //@}
     
     ///@brief Random number distributions
     //@{
     /** Generate a random number in the range [0,1). */
     inline double uniform_01 () {
-        boost::random::uniform_01<pcg32&> rng_uniform01 (m_rng);
+        boost::random::uniform_01<T&> rng_uniform01 (m_rng);
         return rng_uniform01 ();
     }
     
     /** This function returns a Gaussian random variate, with mean mean and
      * standard deviation std. The sampled value x ~ N(mean, std^2) . */
-    double gauss (double mean, double std);
+    double gauss (double mean, double std){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::normal_distribution<> dist (mean, std);
+        return dist(m_rng);
+# else
+        return gsl_ran_gaussian(&m_gsl_gen,std)+mean;
+# endif
+    }
     
     /** This function returns a random variate from the gamma distribution. */
-    double gamma (double a, double b);
+    double gamma (double a, double b){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::gamma_distribution<> dist (a, b);
+        return dist(m_rng);
+# else
+        return gsl_ran_gamma(&m_gsl_gen, a, b);
+# endif
+    }
     
     /** This function returns a random variate from the lognormal distribution.
      * 
@@ -76,7 +160,14 @@ struct RNG {
      * @param mu mean-log
      * @param sigma sigma-log
      */
-    double log_normal (double meanlog, double stdlog);
+    double log_normal (double meanlog, double stdlog){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::lognormal_distribution<> dist (meanlog, stdlog);
+        return dist (m_rng);
+# else
+        return gsl_ran_lognormal (&m_gsl_gen, meanlog, stdlog);
+# endif
+    }
     
     /** Return the maximum over multiple log-normal samples.
      *
@@ -86,10 +177,39 @@ struct RNG {
      * @param meanlog mean of underlying Gaussian
      * @param stdlog standard deviation of underlying Gaussian
      */
-    double max_multi_log_normal (double start, int n, double meanlog, double stdlog);
+    double max_multi_log_normal (double start, int n, double meanlog, double stdlog){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        // We don't have a CDF available, so take log-normal samples
+        double result = start;
+        for (int i = 0; i < n; ++i) {
+            double sample = log_normal(meanlog, stdlog);
+            if (sample > result) result = sample;
+        }
+        return result;
+# else
+        // Used for performance reasons. Calling GSL's log_normal 5 times is 50% slower.
+        // For random variables X1, .., Xn with identical distribution X and
+        // Mn = max(X1, .., Xn), the CDF:
+        //      F_Mn(x) = P(Mn ≤ x)
+        //              = P(X1 ≤ x) P(X2 ≤ x) .. P(Xn ≤ x)
+        //              = F_X(x) ^ n
+        // Thus for u = F_Mn(x) = F_X(x) ^ n, u^(1/n) = F_X(x).
+        double normp = pow( uniform_01(), 1.0 / n );
+        double zval = gsl_cdf_ugaussian_Pinv (normp);
+        double multi_sample = exp(meanlog + stdlog * zval);
+        return std::max(start, multi_sample);
+# endif
+    }
     
     /** This function returns a random variate from the beta distribution. */
-    double beta(double a, double b);
+    double beta(double a, double b){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::beta_distribution<> dist (a, b);
+        return dist(m_rng);
+# else
+        return gsl_ran_beta (&m_gsl_gen,a,b);
+# endif
+    }
     
     /** This function wraps beta(), setting b=b and a such that m is the mean
      * of the distribution. */
@@ -100,11 +220,31 @@ struct RNG {
     }
     
     /** This function returns a random integer from the Poisson distribution with mean lambda. */
-    int poisson(double lambda);
+    int poisson(double lambda){
+        if( !(boost::math::isfinite)(lambda) ){
+            //This would lead to an inifinite loop
+            throw TRACED_EXCEPTION( "lambda is inf", Error::InfLambda );
+        }
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::poisson_distribution<> dist (lambda);
+        return dist(m_rng);
+# else
+        return gsl_ran_poisson (&m_gsl_gen, lambda);
+# endif
+    }
 
     /** This function returns true with probability prob or 0 with probability
      * 1-prob (Bernoulli distribution). */
-    bool bernoulli(double prob);
+    bool bernoulli(double prob){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::bernoulli_distribution<> dist (prob);
+        return dist(m_rng);
+# else
+        assert( (boost::math::isfinite)(prob) );
+        // return true iff our variate is less than the probability
+        return uniform_01() < prob;
+# endif
+    }
     
     /** This function returns an integer from 0 to 1-n, where every value has
      * equal probability of being sampled. */
@@ -121,12 +261,18 @@ struct RNG {
      * @param lambda (λ) is the scale parameter
      * @param k is the shape parameter
      */
-    double weibull( double lambda, double k );
+    double weibull( double lambda, double k ){
+# ifdef OM_RANDOM_USE_BOOST_DIST
+        boost::random::weibull_distribution<> dist (k, lambda);
+        return dist(m_rng);
+# else
+        return gsl_ran_weibull( &m_gsl_gen, lambda, k );
+# endif
+    }
     //@}
     
 private:
-    // I would prefer to use pcg64, but MSVC mysteriously fails
-    pcg32 m_rng;
+    T m_rng;
     
     // Hooks for GSL distributions
     gsl_rng_type m_gsl_type;
@@ -134,7 +280,8 @@ private:
 };
 
 /// The global RNG
-extern RNG global_RNG;
+// I would prefer to use pcg64, but MSVC mysteriously fails
+extern RNG<pcg32> global_RNG;
 
 } }
 #endif
